+++ groovyConsole/scripts/openWithGroovyConsoleFull.groovy
@@ -1,0 +1,4 @@
+import edofro.groovyconsole.GroovyConsole as GC
+
+GC.openGroovyConsole(node , getBinding())
+